:bug: Fix accept_focus behavior on containers
In a container, if `accept_focus` is called, then the focus should
be handled to the widget that is on focus instead of always
searching for the next one. For example, if there are two
containers in a window and the focus is switched from one to
another repeatedly, then the focused child widgets must not change
at all.