:racehorse: Improve compilation time by tweaking anchors
The types related to anchors were highly specialized. Thus,
basically, the functions were being compiled for every call. This
was increasing the compilation time a lot. Now, unfortunately, many
functions related to anchors are type unstable, but this is not as
expressive as the compilation time since usually they are called
only a few times.