:sparkles: Add default signal key_pressed to all widgets
The signal `key_pressed` is emitted every time a key is pressed
when a widget is on focus. In this case, if it returns `nothing`,
then the default action of the widget is performed. It can also
return a boolean value. In this case, no additional action is
performed and the focus is retained only if the returned value is
true.