:warning: Revamp of process focus for widgets
The function `process_focus` is now shared by all widgets. This function
checks if the keystroke is `tab`. If not, then it emits the signal
`key_pressed`. When each widget is created, `key_pressed` is tied to the
default function `process_keystroke`. Hence, the user can now completely
overwrite the default processing algorithm of a widget.